refactor(web-diary): react-app structure refactored
affects: @practice/web-diary, @components/app-header, @components/app, @components/post-add-form,
@components/post-list-item, @components/post-item, @components/filter, @components/search,
@ui/button, @ui/input

ISSUES CLOSED: #56
