--- practice/web-diary/src/components/app/index.tsx
@@ -1,3 +1,0 @@
-import App from './App'
-
-export default App
