+++ practice/web-diary/src/components/app/src/App.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+import '../../../style/app.css'
+import { AppWrapper } from './AppWrapper'
+
+const App = () => {
+  return <AppWrapper />
+}
+
+export default App
