+++ practice/web-diary/src/components/app/src/index.tsx
@@ -1,0 +1,5 @@
+import App from './App'
+import { AppWrapper } from './AppWrapper'
+
+export default App
+export { AppWrapper }
