--- practice/web-diary/src/components/post-add-form/PostAddForm.tsx
@@ -1,20 +1,0 @@
-import React from 'react'
-
-import '../../style/post-add-form.css'
-
-const PostAddForm = () => {
-  return (
-    <form className="bottom-panel d-flex">
-      <input
-        type="text"
-        placeholder="О чем вы думаете сейчас?"
-        className="form-control new-post-label"
-      />
-      <button type="submit" className="btn btn-outline-secondary">
-        Добавить
-      </button>
-    </form>
-  )
-}
-
-export default PostAddForm
