+++ practice/web-diary/src/components/post-add-form/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@components/post-add-form",
+  "version": "0.0.0",
+  "license": "BSD 3-Clause",
+  "main": "src/index.ts",
+  "scripts": {
+    "lint": "eslint"
+  },
+  "dependencies": {
+    "react": "^17.0.1",
+    "@ui/button": "0.0.0",
+    "@ui/input": "0.0.0"
+  }
+}
