+++ practice/web-diary/src/components/post-list-item/src/PostListItem.tsx
@@ -1,17 +1,14 @@
 import React from 'react'
-import '../../style/post-list-item.css'
+import '../../../style/post-list-item.css'
+import { ButtonStar, ButtonTrash } from '../../../ui/button/src'
 
 const PostListItem = () => {
   return (
     <li className="app-list-item d-flex justify-content-between">
       <span className="app-list-item-label">Hello World</span>
       <div className="d-flex justify-content-center align-items-center">
-        <button className="btn-star btn-sm">
-          <i className="fa fa-star"></i>
-        </button>
-        <button className="btn-trash btn-sm">
-          <i className="fa fa-trash-o"></i>
-        </button>
+        <ButtonStar />
+        <ButtonTrash />
         <i className="fa fa-heart"></i>
       </div>
     </li>
