+++ practice/web-diary/src/components/post-list-item/src/index.tsx
