+++ practice/web-diary/src/components/post-list/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@components/post-item",
+  "version": "0.0.0",
+  "license": "BSD 3-Clause",
+  "main": "src/index.ts",
+  "scripts": {
+    "lint": "eslint"
+  },
+  "dependencies": {
+    "react": "^17.0.1",
+    "@components/post-list-item": "0.0.0"
+  }
+}
