--- practice/web-diary/src/components/post-status-filter/PostStatusFilter.tsx
@@ -1,16 +1,0 @@
-import React from 'react'
-
-const PostStatusFilter = () => {
-  return (
-    <div className="btn-group">
-      <button type="button" className="btn btn-info">
-        Все
-      </button>
-      <button type="button" className="btn btn-outline-secondary">
-        Понравилось
-      </button>
-    </div>
-  )
-}
-
-export default PostStatusFilter
