+++ practice/web-diary/src/components/post-status-filter/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@components/filter",
+  "version": "0.0.0",
+  "license": "BSD 3-Clause",
+  "main": "src/index.ts",
+  "scripts": {
+    "lint": "eslint"
+  },
+  "dependencies": {
+    "react": "^17.0.1",
+    "@ui/button": "0.0.0"
+  }
+}
