+++ practice/web-diary/src/components/post-status-filter/src/PostStatusFilter.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import ButtonInfo, { ButtonSecondary } from '../../../ui/button/src'
+
+const PostStatusFilter = () => {
+  return (
+    <div className="btn-group">
+      <ButtonInfo />
+      <ButtonSecondary />
+    </div>
+  )
+}
+
+export default PostStatusFilter
