+++ practice/web-diary/src/components/search-panel/src/SearchPanel.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+import '../../../style/search-panel.css'
+
+import { InputSearch } from '../../../ui/input/src/InputSearch'
+
+const SearchPanel = () => {
+  return <InputSearch />
+}
+
+export default SearchPanel
