+++ practice/web-diary/src/ui/button/src/ButtonInfo.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const ButtonInfo = () => {
+  return (
+    <button type="button" className="btn btn-info">
+      Все
+    </button>
+  )
+}
