+++ practice/web-diary/src/ui/button/src/ButtonStar.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const ButtonStar = () => {
+  return (
+    <button className="btn-star btn-sm">
+      <i className="fa fa-star"></i>
+    </button>
+  )
+}
