+++ practice/web-diary/src/ui/button/src/ButtonSubmit.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const ButtonSubmit = () => {
+  return (
+    <button type="submit" className="btn btn-outline-secondary">
+      Добавить
+    </button>
+  )
+}
