+++ practice/web-diary/src/ui/button/src/ButtonTrash.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const ButtonTrash = () => {
+  return (
+    <button className="btn-trash btn-sm">
+      <i className="fa fa-trash-o"></i>
+    </button>
+  )
+}
