+++ practice/web-diary/src/ui/button/src/ButtonsSecondary.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const ButtonSecondary = () => {
+  return (
+    <button type="button" className="btn btn-outline-secondary">
+      Понравилось
+    </button>
+  )
+}
