+++ practice/web-diary/src/ui/button/src/index.tsx
@@ -1,0 +1,11 @@
+import { ButtonInfo } from './ButtonInfo'
+import { ButtonSecondary } from './ButtonsSecondary'
+import { ButtonStar } from './ButtonStar'
+import { ButtonTrash } from './ButtonTrash'
+import { ButtonSubmit } from './ButtonSubmit'
+
+export default ButtonInfo
+export { ButtonSecondary }
+export { ButtonStar }
+export { ButtonTrash }
+export { ButtonSubmit }
