+++ practice/web-diary/src/ui/input/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@ui/input",
+  "version": "0.0.0",
+  "license": "BSD 3-Clause",
+  "main": "src/index.ts",
+  "scripts": {
+    "lint": "eslint"
+  },
+  "dependencies": {
+    "react": "^17.0.1"
+  }
+}
