+++ practice/web-diary/src/ui/input/src/InputNewForm.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export const InputNewForm = () => {
+  return (
+    <input
+      type="text"
+      placeholder="О чем вы думаете сейчас?"
+      className="form-control new-post-label"
+    />
+  )
+}
