+++ practice/web-diary/src/ui/input/src/index.tsx
@@ -1,0 +1,5 @@
+import { InputSearch } from './InputSearch'
+import { InputNewForm } from './InputNewForm'
+
+export default InputSearch
+export { InputNewForm }
